# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.DetailInformationRes:
    properties:
      booking_id:
        type: integer
      category:
        type: string
      due_date:
        type: string
      movie_title:
        type: string
      payment_status:
        type: string
      qr_code:
        type: string
      quantity:
        type: integer
      seats:
        items:
          type: string
        type: array
      showtime_date:
        type: string
      showtime_time:
        type: string
      ticket_status:
        type: string
      total_price:
        type: integer
      virtua_account:
        type: integer
    type: object
...
  dto.OrderHistoryRes:
    properties:
      booking_id:
        type: integer
      cinema_logo:
        type: string
      cinema_name:
        type: string
      movie_title:
        type: string
      showtime:
        type: string
      status_paid:
        type: string
      status_ticket:
        type: string
    type: object
...
  dto.ResponseError:
    properties:
      error:
        example: internal server error/bad request
        type: string
      message:
        example: Failed get data/internal server error
        type: string
      status:
        example: error
        type: string
    type: object
...
  dto.UserProfileResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      full_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      location:
        type: string
      phone:
        type: string
      photo:
        type: string
      point:
        type: integer
    type: object
  dto.UserUpdateProfileRes:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
      photo:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: limustadji@gmail.com
    name: alpha male
    url: https://github.com/L1mus
  description: This is a backend project for a web application called E-wallet.
  license:
    name: MIT
  title: Movie API
  version: "1.0"
paths:
...
  /users/history:
    get:
      consumes:
      - application/json
      description: Get list of Order History for logged-in User
      produces:
      - application/json
      responses:
        "200":
          description: Get Order History Succesfully
          schema:
            items:
              $ref: '#/definitions/dto.OrderHistoryRes'
            type: array
        "401":
          description: 'Unauthorized: Token not exist / Format token invalid'
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get User Order History
      tags:
      - Users
  /users/history/{id}/detail:
    get:
      consumes:
      - application/json
      description: Get detailed information for specific order history by user logged-in
      parameters:
      - description: Booking Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get Detail Information User Succesfully
          schema:
            $ref: '#/definitions/dto.DetailInformationRes'
        "400":
          description: Invalid Booking Id Format
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: 'Unauthorized: Token not exist / Format token invalid'
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get Detail Information
      tags:
      - Users
  /users/profile:
    get:
      consumes:
      - application/json
      description: Get the profile data of the logged in user based on the jwt token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserProfileResponse'
        "401":
          description: 'Unauthorized: Token not exist'
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get User Profile
      tags:
      - Users
    patch:
      consumes:
      - multipart/form-data
      description: Update user data profile
      parameters:
      - description: Update First Name
        in: formData
        name: first_name
        type: string
      - description: Update Last Name
        in: formData
        name: last_name
        type: string
      - description: Update Phone Number
        in: formData
        name: phone
        type: string
      - description: Update New Password
        in: formData
        name: new_password
        type: string
      - description: Confirm New Password
        in: formData
        name: confirm_password
        type: string
      - description: Update Photo Profile
        in: formData
        name: photo
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Update Profile Succesfully
          schema:
            $ref: '#/definitions/dto.UserUpdateProfileRes'
        "400":
          description: Invalid Input Data
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "401":
          description: 'Unauthorized: Token not exist / Format token invalid'
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "422":
          description: File Too Large / Invalid File Format
          schema:
            $ref: '#/definitions/dto.ResponseError'
        "500":
          description: Failed to Update Profile / Failed to Save Image
          schema:
            $ref: '#/definitions/dto.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update User Profile
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    description: Bearer token used for authorization
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"